requirements.txt
Flask==3.1.2
gevent==25.5.1
mysql-connector-python==9.4.0
pillow==11.3.0
